[migrations]: Changed migrations price and quantity to 64 bit

+++ migrations/20161205053332_create_users_table.up.sql
@@ -2,8 +2,8 @@ CREATE TABLE IF NOT EXISTS Users (
 	id int(11) UNSIGNED NOT NULL AUTO_INCREMENT,
 	email varchar(255) UNIQUE,
 	name varchar(255) NOT NULL,
-	cash int(11) UNSIGNED NOT NULL,
-	total int(11) SIGNED NOT NULL,
+	cash bigint(11) UNSIGNED NOT NULL,
+	total bigint(11) SIGNED NOT NULL,
 	createdAt varchar(255) NOT NULL DEFAULT "0000-00-00T00:00:00+05:30",
 	PRIMARY KEY (id)
 ) AUTO_INCREMENT=1;
